lua: relax fp signaling configuration
This gives user code a chance to suffer^Whandle out-of-range and invalid
results of floating point operations, rather than having the OS kill the
entire process.

Killing the process may be useful for a raw C program, since its state
can be easily inspected to find the place where the (likely) bug
occured.  This isn't the case with Lua programs, for which the process
state is mostly useless in determining the Lua source line that caused
the VM to execute an invalid fp operation.